ai/refactor: 型定義をinternal/types/types.goに統一し重複を解消
